src/gcp-data-characterizer-llm/requirements.txt
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0
numpy>=1.24.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.0
ydata-profiling>=4.5.0
plotly>=5.18.0
openai>=1.0.0
python-dotenv>=1.0.0
